+++ Packs/Legacy/Integrations/TruSTAR/CHANGELOG.md
@@ -1,6 +1,13 @@
 ## [Unreleased]
--
-
+- Added 3 new commands:
+  - ***trustar-get-phishing-submissions***
+  - ***trustar-get-phishing-indicators***
+  - ***trustar-set-triage-status***
+- Deprecated the following commands:
+  - ***file***
+  - ***url***
+  - ***ip***
+  - ***domain***
 
 ## [20.4.0] - 2020-04-14
 -